refactor(libs-api-services-node): create neo4j schema service
refactor(libs-api-services-node): create neo4j schema service

refactor(libs-api-services-node): create neo4j schema service
