+++ libs/api/services/node/src/index.ts
@@ -1,1 +1,2 @@
 export * from './neo4j'
+export * from './graphql'
