--- libs/api/services/node/src/neo4j/database/index.ts
@@ -1,2 +1,0 @@
-export * from './neo4j-database.module'
-export * from './neo4j-database.provider'
