--- libs/api/services/node/src/neo4j/database/neo4j-database.module.ts
@@ -1,11 +1,0 @@
-import { Module } from '@nestjs/common'
-import {
-  NEO4J_DATABASE_PROVIDER,
-  neo4DatabaseProvider,
-} from './neo4j-database.provider'
-
-@Module({
-  providers: [neo4DatabaseProvider],
-  exports: [NEO4J_DATABASE_PROVIDER],
-})
-export class Neo4jDatabaseModule {}
