+++ libs/api/services/node/src/neo4j/index.ts
@@ -1,1 +1,4 @@
 export * from './neo4j-node.module'
+export * from './neo4j.provider'
+export * from './neo4j-node.service'
+export * from './neo4j-schema.service'
